Conflicts:
	oneflow/core/boxing/eager_boxing_interpreter_mgr.cpp
	oneflow/core/functional/functional_api.yaml
	oneflow/core/functional/impl/slice_boxing_functor.cpp
	oneflow/user/kernels/communicate_util.cpp
	oneflow/user/kernels/communicate_util.h
	python/oneflow/test/modules/test_eager_boxing.py

+++ oneflow/user/ops/argmax_op.cpp
@@ -34,7 +34,7 @@ REGISTER_NO_GRAD_USER_OP("argmax")
       return Maybe<void>::Ok();
     })
     .SetDataTypeInferFn([](user_op::InferContext* ctx) -> Maybe<void> {
-      *ctx->OutputDType("out", 0) = DataType::kInt32;
+      *ctx->OutputDType("out", 0) = DataType::kInt64;
       return Maybe<void>::Ok();
     });
